Added phone number editing.
Because a design goal of this project is to heavily
rely on vanilla javascript, this section is a bit
verbose due to the lack of a templating engine.

However, performance is very snappy and all of this
code is fairly well self-contained.

Also replaced classic-functions with arrow-functions.
This makes the code base incompatible with older
browsers (specifically IE 11), so I also removed
the polyfills.