Formatting (CSS) tweaks.
Fix a bullet point render issue.
The "li li:before" rule used the property "content" to display a unicode square.
That rendered too large on iOS.
It also did not vertically-center properly with varying font sizes.
Using the rule "ul ul" with the property "list-style-type" works well.

Added a media query for line-height.
The value of "2" was too large for small mobile screens, but it works well for large screens.